third_party/WebKit/LayoutTests/http/tests/inspector/network/resource-priority-expected.txt

CONSOLE WARNING: line 22: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Tests resource priorities.


Request: empty-script.js?sync priority: High
Request: empty-script.js?async priority: Low
Request: abe.png?preceding priority: Low
Request: empty-script.js?preceded priority: Medium
Request: abe.png?precedingPreload priority: Low
Request: empty-script.js?precededByPreload priority: High
Request: empty.html?xhr-sync priority: VeryHigh
Request: empty.html?xhr-async priority: High
Request: abe.png priority: Low
Request: style.css priority: VeryHigh
Request: empty.html?iframe priority: VeryHigh
